feat : ProductController, Repository, Service 예상 보험료 계산 로직 수정(타입 Double->Integer 변환)

+++ src/main/java/com/example/vvs/domain/product/repository/ProductRepository.java
@@ -10,5 +10,5 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
     List<Product> findAllByOrderByIdDesc();
     List<Product> findAllByCategoryOrderByIdDesc(String category);
     @Query(value = "select price from Product  where id= :id")
-    Double findPriceById(Long id);
+    Integer findPriceById(Long id);
 }
